moved Read_metadata_thread class out of build_metadata_db.py so it can also be called from test_build_metadata_db.py for reading a single file, also improved printing in db_functions.py
